Splitted options and legacy code from Connection.
