+++ ts/Gantt/ConnectorsDefaults.ts
@@ -1,0 +1,319 @@
+/* *
+ *
+ *  (c) 2016 Highsoft AS
+ *  Authors: Øystein Moseng, Lars A. V. Cabrera
+ *
+ *  License: www.highcharts.com/license
+ *
+ *  !!!!!!! SOURCE GETS TRANSPILED BY TYPESCRIPT. EDIT TS FILE ONLY. !!!!!!!
+ *
+ * */
+
+'use strict';
+
+/* *
+ *
+ *  Imports
+ *
+ * */
+
+import type ConnectorsOptions from './ConnectorsOptions';
+
+/* *
+ *
+ *  Declarations
+ *
+ * */
+
+export interface ConnectorsDefaults {
+    connectors: ConnectorsOptions;
+}
+
+/* *
+ *
+ *  API Defaults
+ *
+ * */
+
+const connectorsDefaults: ConnectorsDefaults = {
+    /**
+     * The Pathfinder module allows you to define connections between any two
+     * points, represented as lines - optionally with markers for the start
+     * and/or end points. Multiple algorithms are available for calculating how
+     * the connecting lines are drawn.
+     *
+     * Connector functionality requires Highcharts Gantt to be loaded. In Gantt
+     * charts, the connectors are used to draw dependencies between tasks.
+     *
+     * @see [dependency](series.gantt.data.dependency)
+     *
+     * @sample gantt/pathfinder/demo
+     *         Pathfinder connections
+     *
+     * @declare      Highcharts.ConnectorsOptions
+     * @product      gantt
+     * @optionparent connectors
+     */
+    connectors: {
+
+        /**
+         * Enable connectors for this chart. Requires Highcharts Gantt.
+         *
+         * @type      {boolean}
+         * @default   true
+         * @since     6.2.0
+         * @apioption connectors.enabled
+         */
+
+        /**
+         * Set the default dash style for this chart's connecting lines.
+         *
+         * @type      {string}
+         * @default   solid
+         * @since     6.2.0
+         * @apioption connectors.dashStyle
+         */
+
+        /**
+         * Set the default color for this chart's Pathfinder connecting lines.
+         * Defaults to the color of the point being connected.
+         *
+         * @type      {Highcharts.ColorString}
+         * @since     6.2.0
+         * @apioption connectors.lineColor
+         */
+
+        /**
+         * Set the default pathfinder margin to use, in pixels. Some Pathfinder
+         * algorithms attempt to avoid obstacles, such as other points in the
+         * chart. These algorithms use this margin to determine how close lines
+         * can be to an obstacle. The default is to compute this automatically
+         * from the size of the obstacles in the chart.
+         *
+         * To draw connecting lines close to existing points, set this to a low
+         * number. For more space around existing points, set this number
+         * higher.
+         *
+         * @sample gantt/pathfinder/algorithm-margin
+         *         Small algorithmMargin
+         *
+         * @type      {number}
+         * @since     6.2.0
+         * @apioption connectors.algorithmMargin
+         */
+
+        /**
+         * Set the default pathfinder algorithm to use for this chart. It is
+         * possible to define your own algorithms by adding them to the
+         * Highcharts.Pathfinder.prototype.algorithms object before the chart
+         * has been created.
+         *
+         * The default algorithms are as follows:
+         *
+         * `straight`:      Draws a straight line between the connecting
+         *                  points. Does not avoid other points when drawing.
+         *
+         * `simpleConnect`: Finds a path between the points using right angles
+         *                  only. Takes only starting/ending points into
+         *                  account, and will not avoid other points.
+         *
+         * `fastAvoid`:     Finds a path between the points using right angles
+         *                  only. Will attempt to avoid other points, but its
+         *                  focus is performance over accuracy. Works well with
+         *                  less dense datasets.
+         *
+         * Default value: `straight` is used as default for most series types,
+         * while `simpleConnect` is used as default for Gantt series, to show
+         * dependencies between points.
+         *
+         * @sample gantt/pathfinder/demo
+         *         Different types used
+         *
+         * @type    {Highcharts.PathfinderTypeValue}
+         * @default undefined
+         * @since   6.2.0
+         */
+        type: 'straight',
+
+        /**
+         * The corner radius for this chart's Pathfinder connecting lines
+         *
+         * @since next
+         */
+        radius: 0,
+
+        /**
+         * Set the default pixel width for this chart's Pathfinder connecting
+         * lines.
+         *
+         * @since 6.2.0
+         */
+        lineWidth: 1,
+
+        /**
+         * Marker options for this chart's Pathfinder connectors. Note that
+         * this option is overridden by the `startMarker` and `endMarker`
+         * options.
+         *
+         * @declare Highcharts.ConnectorsMarkerOptions
+         * @since   6.2.0
+         */
+        marker: {
+            /**
+             * Set the radius of the connector markers. The default is
+             * automatically computed based on the algorithmMargin setting.
+             *
+             * Setting marker.width and marker.height will override this
+             * setting.
+             *
+             * @type      {number}
+             * @since     6.2.0
+             * @apioption connectors.marker.radius
+             */
+
+            /**
+             * Set the width of the connector markers. If not supplied, this
+             * is inferred from the marker radius.
+             *
+             * @type      {number}
+             * @since     6.2.0
+             * @apioption connectors.marker.width
+             */
+
+            /**
+             * Set the height of the connector markers. If not supplied, this
+             * is inferred from the marker radius.
+             *
+             * @type      {number}
+             * @since     6.2.0
+             * @apioption connectors.marker.height
+             */
+
+            /**
+             * Set the color of the connector markers. By default this is the
+             * same as the connector color.
+             *
+             * @type      {Highcharts.ColorString|Highcharts.GradientColorObject|Highcharts.PatternObject}
+             * @since     6.2.0
+             * @apioption connectors.marker.color
+             */
+
+            /**
+             * Set the line/border color of the connector markers. By default
+             * this is the same as the marker color.
+             *
+             * @type      {Highcharts.ColorString}
+             * @since     6.2.0
+             * @apioption connectors.marker.lineColor
+             */
+
+            /**
+             * Enable markers for the connectors.
+             */
+            enabled: false,
+
+            /**
+             * Horizontal alignment of the markers relative to the points.
+             *
+             * @type {Highcharts.AlignValue}
+             */
+            align: 'center',
+
+            /**
+             * Vertical alignment of the markers relative to the points.
+             *
+             * @type {Highcharts.VerticalAlignValue}
+             */
+            verticalAlign: 'middle',
+
+            /**
+             * Whether or not to draw the markers inside the points.
+             */
+            inside: false,
+
+            /**
+             * Set the line/border width of the pathfinder markers.
+             */
+            lineWidth: 1
+        },
+
+        /**
+         * Marker options specific to the start markers for this chart's
+         * Pathfinder connectors. Overrides the generic marker options.
+         *
+         * @declare Highcharts.ConnectorsStartMarkerOptions
+         * @extends connectors.marker
+         * @since   6.2.0
+         */
+        startMarker: {
+            /**
+             * Set the symbol of the connector start markers.
+             */
+            symbol: 'diamond'
+        },
+
+        /**
+         * Marker options specific to the end markers for this chart's
+         * Pathfinder connectors. Overrides the generic marker options.
+         *
+         * @declare Highcharts.ConnectorsEndMarkerOptions
+         * @extends connectors.marker
+         * @since   6.2.0
+         */
+        endMarker: {
+            /**
+             * Set the symbol of the connector end markers.
+             */
+            symbol: 'arrow-filled'
+        }
+    }
+};
+
+/**
+ * Override Pathfinder connector options for a series. Requires Highcharts Gantt
+ * to be loaded.
+ *
+ * @declare   Highcharts.SeriesConnectorsOptionsObject
+ * @extends   connectors
+ * @since     6.2.0
+ * @excluding enabled, algorithmMargin
+ * @product   gantt
+ * @apioption plotOptions.series.connectors
+ */
+
+/**
+ * Connect to a point. This option can be either a string, referring to the ID
+ * of another point, or an object, or an array of either. If the option is an
+ * array, each element defines a connection.
+ *
+ * @sample gantt/pathfinder/demo
+ *         Different connection types
+ *
+ * @declare   Highcharts.XrangePointConnectorsOptionsObject
+ * @type      {string|Array<string|*>|*}
+ * @extends   plotOptions.series.connectors
+ * @since     6.2.0
+ * @excluding enabled
+ * @product   gantt
+ * @requires  highcharts-gantt
+ * @apioption series.xrange.data.connect
+ */
+
+/**
+ * The ID of the point to connect to.
+ *
+ * @type      {string}
+ * @since     6.2.0
+ * @product   gantt
+ * @apioption series.xrange.data.connect.to
+ */
+
+''; // detach doclets above
+
+/* *
+ *
+ *  Default Export
+ *
+ * */
+
+export default connectorsDefaults;
