+++ ts/Series/Gantt/GanttSeriesOptions.d.ts
@@ -17,6 +17,10 @@
  * */
 
 import type AnimationOptions from '../../Core/Animation/AnimationOptions';
+import type {
+    ConnectorsStartMarkerOptions,
+    ConnectorsOptions
+} from '../../Gantt/ConnectorsOptions';
 import type ColorType from '../../Core/Color/ColorType';
 import type GanttSeries from './GanttSeries';
 import type { SeriesStatesOptions } from '../../Core/Series/SeriesOptions';
@@ -32,12 +36,12 @@ export interface GanttAnimationOptions extends Partial<AnimationOptions> {
     reversed?: boolean;
 }
 
-export interface GanttConnectorOptions extends Highcharts.ConnectorsOptions {
+export interface GanttConnectorOptions extends ConnectorsOptions {
     animation?: (boolean|GanttAnimationOptions);
     startMarker?: GanttConnectorStartMarkerOptions;
 }
 
-export interface GanttConnectorStartMarkerOptions extends Highcharts.ConnectorsStartMarkerOptions {
+export interface GanttConnectorStartMarkerOptions extends ConnectorsStartMarkerOptions {
     fill: ColorType;
 }
